### 2021-07-14 #3493 @whedon
[PRE REVIEW]: BisPy: Bisimulation in Python
**Submitting author:** @fAndreuzzi (<a href="http://orcid.org/0000-0002-9508-7801">Francesco Andreuzzi</a>)
**Repository:** <a href="https://github.com/fAndreuzzi/BisPy" target ="_blank">https://github.com/fAndreuzzi/BisPy</a>
**Version:** 0.2.0
