### 2021-07-19 #3493 @mjsottile
Hi @fAndreuzzi  : I have found one reviewer via email and will reach out to others on here to find one or two more.  In the meantime, I looked at the paper and it has a few typos (e.g., line 41 - I assume "parse" should be "sparse"?).  I would recommend doing a proofreading pass to find simple errors like that.  It might also be worth adding a small example in the paper that demonstrates the package so a reader can get a sense for it without having to go look at code right away.

I expect I will be able to start the formal review soon once I have the additional reviewer(s).
