### 2021-07-19 #3493 @fAndreuzzi
> I would recommend doing a proofreading pass to find simple errors like that.

Thank you, I will review the paper again to fix this kind of typo.

>  It might also be worth adding a small example

Sure, I'm going to add a meaningful example.
