### 2021-07-30 #3519 @danielskatz
@mjsottile - please use @openjournals/dev for technical problems, and `@openjournals/joss-eics` for the editors in chief for other things, in case one person, for example @arfon, is on vacation.  And I'll check on this now
