### 2021-07-30 #3519 @danielskatz
Ok, this is now fine, as seen in the first comment in this issue, where it now says `Reviewers: @jonjoncardoso, @mschordan, @zoometh`  When I started, zoometh was listed twice, since you had added them twice, so I removed them then readded them so they would only be listed once.  The whedon command also shows this correctly in its response immediately above, but this is less important than what the first comment shows.  Now you have to manually edit the first comment and add a checklist for zoometh, as whedon only generates these checklists once, when the review issue is created, based on the reviewers that were assigned at the time (in the pre-review issue). If you need help with this, let `@openjournals/joss-eics` know.
