### 2021-07-30 #3519 @mjsottile
@danielskatz thanks.  When I looked at it last night the reviewer did show up in the issue body and had a checklist, but was missing from the list of assignees for the GitHub issue itself.  They appear to be listed as an assignee now.
