### 2021-08-01 #3519 @zoometh
@fAndreuzzi I'll suggest reordering the parts of your paper & to modify some:

1. **Summary**: make it shorter, currently it is too specific (ie not enough general)
2. **Statement of need**: has to be placed _before_ the **Examples**
3. ~ **BisPy framework**: present your Python package here
4. **Examples**
5. ...

I've also noted:
- There is a missing [community guideline](https://joss.readthedocs.io/en/latest/review_checklist.html#documentation) for contributions' Code of conduct
- Few typo errors:
  -  in [README](https://github.com/fAndreuzzi/BisPy#dependencies-and-installation): "_requires requires_"
  - I'm not sure to understand this sentence "_Again, the two nodes are almost indistinguishable: we may switch `b` and `b`, and unless `a` resents the node `c` for some reason nobody would notice_" in the [docs/](https://github.com/fAndreuzzi/BisPy/blob/master/docs/source/index.rst), maybe a typo error
