### 2021-08-01 #3519 @fAndreuzzi
Hi @zoometh, thank you for your feedback.

I worked a little bit on the README (fixed the typo you mentioned, and added a [new](https://github.com/fAndreuzzi/BisPy/blob/master/README.md#reporting-a-bug) section to explain in detail how to report a bug. In [this](https://github.com/fAndreuzzi/BisPy#how-to-contribute) section I explain how to submit a pull request, is this enough for you?

I'm going to work on the paper to apply your suggestions, I'll update you here when I'm satisfied with the changes.
