### 2021-08-05 #3519 @zoometh
>
>
> @zoometh In the last version I added the section **BisPy** and slightly modified the **Statement of need** accordingly. What do you think?

@fAndreuzzi  All right, that's fine for me. Great job.
