### 2021-08-19 #3519 @fAndreuzzi
Thank you very much @mschordan, the right output is indeed the one given by **BisPy**, I probably confused the outputs while doing copy-paste from the terminal to the browser. I fixed `README.md` in my last commit.
