### 2021-08-20 #3519 @jonjoncardoso
@fAndreuzzi I haven't gone through the code yet but some general comments so far:

- I know you have pointed to reference elsewhere but I feel like some references are also needed to support this sentence about the problem complexity: _"[...] several algorithms for the computation of maximum bisimulation have been studied throughout the years, and it is now widely known that the problem has an O(|E| log |V |) algorithmic solution"_ (Line 15).
- I suggest that recent changes you made to the summary and README should also be replicated to your readthedocs page. The description of bisimulation problem is clearer on the paper manuscript version compared to the page, for instance.
