### 2021-08-20 #3519 @fAndreuzzi
> some references are also needed to support this sentence about the problem complexity

Sure, I'm still not completely familiar with the rule "cite whenever you say something which is not obvious, I fixed the problem in one of the latest commits.

> The description of bisimulation problem is clearer on the paper manuscript version

Could you please point me the parts where I explained the bisimulation better in the paper rather than in the documentation? In my opinion the explanation in the paper is very limited. Anyway I improved a little bit the explanation on [ReadTheDocs](https://bispy-bisimulation-in-python.readthedocs.io/en/latest/) (cleared some sentences and added a new paragraph about maximum bisimulation). Kindly let me know what you think.

Thank you
