### 2021-09-07 #3519 @mschordan
@fAndreuzzi
1) You write: "We found some sparse implementations of the most famous algorithm for the computation of the maximum bisimulation (Paige-Tarjan), however the source code lacked documentation and test cases, and did not seem to be intended for distribution."

This may sound strange to a reader. How did you find it, when it was not intended for distribution? Can you provide a reference? If not, it's not clear why this is mentioned in the first place.

2) You write: "Other algorithms to compute the maximum bisimulation are available, but for what we could see they are slight variations of one of those mentioned above, or tailored on particular cases."

Which algorithms are you referring to? What are "slight variations"? Any algorithms that are referred to should be cited properly.

3) You write: "Our implementations have been tested and documented deeply;" - That's indeed an important aspect as you mention also that "our package may be useful to assess the performance". Some more information to make this point could be interesting to a potential user.
What was the biggest input that you tested the algorithm on? Can you provide some runtime measurement?
