### 2021-09-11 #3519 @fAndreuzzi
Hi @mschordan, sorry for the delay but I was a little bit busy this week.

> How did you find it, when it was not intended for distribution?

I found code snippets on GitHub, but as I mention in the paper they lacked documentation, test cases, and a brief examination of the code made me think that time complexity was not taken into account at all. When I say "_not intended for distribution_" I mean that those snippets are not meant to be included in someone else's code or used for experiments whose intent is something more than "_what is the maximum bisimulation of this graph?_". Maybe I should change "_distribution_" to something else in order to express the meaning in a better way?

> Which algorithms are you referring to? What are "slight variations"? Any algorithms that are referred to should be cited properly.

If this sentence causes problems I think I can remove it from the paper. I do not think that it is interesting or meaningful to cite algorithms which are not those that I implemented.

> What was the biggest input that you tested the algorithm on? Can you provide some runtime measurement?

Sure, do you think I should rather include runtime measurement in `readme.md` or in the paper?
