### 2021-09-15 #3519 @mschordan
@fAndreuzzi
> > How did you find it, when it was not intended for distribution?
>
> I found code snippets on GitHub, but as I mentioned in the paper they lacked documentation, test cases, and a brief examination of the code made me think that time complexity was not taken into account at all. When I say "_not intended for distribution_" I mean that those snippets are not meant to be included in someone else's code or used for experiments whose intent is something more than "_what is the maximum bisimulation of this graph?_". Maybe I should change "_distribution_" to something else in order to express the meaning in a better way?

If you want to mention those, then there should also be a link to the respective GitHub repo, otherwise it's not clear what you are referring to. Based on whether you want to provide a reference/link you can decide whether you want to mention it.

>
> > Which algorithms are you referring to? What are "slight variations"? Any algorithms that are referred to should be cited properly.
>
> If this sentence causes problems I think I can remove it from the paper. I do not think that it is interesting or meaningful to cite algorithms which are not those that I implemented.

Indeed.

> > What was the biggest input that you tested the algorithm on? Can you provide some runtime measurement?
>
> Sure, do you think I should rather include runtime measurement in `readme.md` or in the paper?

Where you mention that "our package may be useful to assess the performance", there it would be good to also provide some numbers on the runtime for a given larger input (and which machine it was run on). This would be good to be mentioned in the paper. Also when you mention "tested deeply", it's better to provide some actual info, e.g. how many tests you provide and what your biggest test input was, how the tests were selected, etc.
