### 2021-09-25 #3519 @fAndreuzzi
Hi @mjsottile, I would like to ask if you could consider the possibility of considering the review over even though there are several unticked checkboxes. I see this happened some times on JOSS, and that the average number of reviewers per paper is usually 2.

I'm proposing this solution since we have no information about a date in which the last reviewer might have some free time to complete the review.
