### 2021-09-25 #3519 @mjsottile
Hi @fAndreuzzi : that is my plan.  When a reviewer disappears with a partial review, but we have other completed reviews, I usually give them a few days after my last ping to see if they return.  Otherwise I carry on without it.  I believe we have enough with the completed reviews - I just need to give it a day or so until I have time to do my own final pass over the completed reviews and the paper itself.  I expect to have time to get to that by the end of the weekend or early in the week.
