### 2021-09-27 #3519 @mjsottile
Excellent.  @fAndreuzzi - I have done a pass over the paper and have found a couple of minor grammatical mistakes to correct:

- line 43-44 : "moreover we splitted the algorithms in smaller functions" should read "moreover we split the algorithms into smaller functions".
- line 86 : "taken by step 4. on a sample of" -- remove the period after the digit 4.
- line 88 : "interests" should be "interest"
- line 90 : "to remark this improvement".  I believe the word "remark" should be replaced with the word "illustrate".
- line 92 : "operative system" should be "operating system"

Once you've finished making these corrections, I can proceed to the final steps.

The final step in the process before I hand it off to the EIC team is to make a tagged release and archive. JOSS needs the software to be archived in an archival repository, such as Zenodo, figshare, or an institutional repository. Once this is done, that repository should provide a DOI to the archive. Please post the DOI that you receive in this thread so I can add it to the submission. Make sure that the version of the software on this review issue matches the version of the software that you archive - let me know if I need to update it.

Once you have completed this, I'll run the final steps and hand the submission off to get it finished up.
