### 2021-09-28 #3519 @danielskatz
@fAndreuzzi - I've suggested some small changes in the paper and bib in https://github.com/fAndreuzzi/BisPy/pull/36 - please merge this, or let me know what you disagree with, then we can proceed to acceptance and publishing.
